Move the json perftests utility to a separate header
