+++ libs/core/testing/CMakeLists.txt
@@ -7,7 +7,7 @@
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
 # Default location is $HPX_ROOT/libs/testing/include
-set(testing_headers hpx/modules/testing.hpp)
+set(testing_headers hpx/modules/testing.hpp hpx/testing/performance.hpp)
 
 # Default location is $HPX_ROOT/libs/testing/include_compatibility
 # cmake-format: off
@@ -18,7 +18,7 @@ set(testing_compat_headers
 # cmake-format: on
 
 # Default location is $HPX_ROOT/libs/testing/src
-set(testing_sources testing.cpp)
+set(testing_sources testing.cpp performance.cpp)
 
 include(HPX_AddModule)
 add_hpx_module(
