+++ libs/core/testing/include/hpx/modules/testing.hpp
@@ -1,5 +1,4 @@
 ////////////////////////////////////////////////////////////////////////////////
-//  Copyright (c) 2020 ETH Zurich
 //  Copyright (c) 2011 Bryce Lelbach
 //
 //  SPDX-License-Identifier: BSL-1.0
@@ -16,18 +15,15 @@
 #include <hpx/preprocessor/expand.hpp>
 #include <hpx/preprocessor/nargs.hpp>
 #include <hpx/preprocessor/stringize.hpp>
+#include <hpx/testing/performance.hpp>
 #include <hpx/thread_support/spinlock.hpp>
 #include <hpx/util/ios_flags_saver.hpp>
 
 #include <atomic>
-#include <chrono>
 #include <cstddef>
 #include <cstdint>
-#include <map>
 #include <mutex>
 #include <ostream>
-#include <string>
-#include <vector>
 
 namespace hpx { namespace util {
 
@@ -179,64 +175,6 @@ namespace hpx { namespace util {
 
         HPX_CORE_EXPORT extern fixture global_fixture;
 
-        // Json output for performance reports
-        class json_perf_times
-        {
-            using key_t = std::tuple<std::string, std::string>;
-            using value_t = std::vector<double>;
-            using map_t = std::map<key_t, value_t>;
-
-            map_t m_map;
-
-            friend std::ostream& operator<<(
-                std::ostream& strm, json_perf_times const& obj)
-            {
-                strm << "{\n";
-                strm << "  \"outputs\" : [";
-                int outputs = 0;
-                for (auto&& item : obj.m_map)
-                {
-                    if (outputs)
-                        strm << ",";
-                    strm << "\n    {\n";
-                    strm << "      \"name\" : \"" << std::get<0>(item.first)
-                         << "\",\n";
-                    strm << "      \"executor\" : \"" << std::get<1>(item.first)
-                         << "\",\n";
-                    strm << "      \"series\" : [";
-                    int series = 0;
-                    for (auto val : item.second)
-                    {
-                        if (series)
-                            strm << ", ";
-                        strm << val;
-                        ++series;
-                    }
-                    strm << "]\n";
-                    strm << "    }";
-                    ++outputs;
-                }
-                if (outputs)
-                    strm << "\n  ";
-                strm << "]\n";
-                strm << "}\n";
-                return strm;
-            }
-
-        public:
-            void add(std::string const& name, std::string const& executor,
-                double time)
-            {
-                m_map[key_t(name, executor)].push_back(time);
-            }
-        };
-
-        json_perf_times& times();
-
-        // Add time to the map for performance report
-        void add_time(std::string const& test_name, std::string const& executor,
-            double time);
-
     }    // namespace detail
 
     ////////////////////////////////////////////////////////////////////////////
@@ -246,9 +184,6 @@ namespace hpx { namespace util {
     HPX_CORE_EXPORT void print_cdash_timing(
         const char* name, std::uint64_t time);
 
-    HPX_CORE_EXPORT void perf_test_report(std::string const& name,
-        std::string const& exec, const std::size_t steps,
-        function_nonser<void(void)>&& test);
 }}    // namespace hpx::util
 
 ////////////////////////////////////////////////////////////////////////////////
