+++ libs/core/testing/src/testing.cpp
@@ -19,7 +19,6 @@
 #include <functional>
 #include <iostream>
 #include <string>
-#include <type_traits>
 
 namespace hpx { namespace util {
     static test_failure_handler_type test_failure_handler;
@@ -72,18 +71,6 @@ namespace hpx { namespace util {
 
         fixture global_fixture{std::cerr};
 
-        json_perf_times& times()
-        {
-            static json_perf_times res;
-            return res;
-        }
-
-        void add_time(std::string const& test_name, std::string const& executor,
-            double time)
-        {
-            times().add(test_name, executor, time);
-        }
-
     }    // namespace detail
 
     ////////////////////////////////////////////////////////////////////////////
@@ -122,30 +109,6 @@ namespace hpx { namespace util {
     void print_cdash_timing(const char* name, std::uint64_t time)
     {
         print_cdash_timing(name, time / 1e9);
-    }
-
-    void perf_test_report(std::string const& name, std::string const& exec,
-        const std::size_t steps, function_nonser<void(void)>&& test)
-    {
-        if (steps == 0)
-            return;
-        // First iteration to cache the data
-        test();
-        using timer = std::chrono::high_resolution_clock;
-        timer::time_point start;
-        for (size_t i = 0; i != steps; ++i)
-        {
-            // For now we don't flush the cache
-            //flush_cache();
-            start = timer::now();
-            test();
-            // default is in seconds
-            auto time =
-                std::chrono::duration_cast<std::chrono::duration<double>>(
-                    timer::now() - start);
-            detail::add_time(name, exec, time.count());
-        }
-        std::cout << detail::times();
     }
 
 }}    // namespace hpx::util
